Added exception folder

+++ src/main/java/com/projectthree/springbanking/transactions/TransactionsController.java
@@ -1,8 +1,12 @@
 package com.projectthree.springbanking.transactions;
 
 import java.util.List;
+import java.util.NoSuchElementException;
 
+import com.projectthree.springbanking.exception.SpringBankingAPIException;
+import com.projectthree.springbanking.exception.SpringBankingServerException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -13,6 +17,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.projectthree.springbanking.accounts.AccountsEntity;
+
+import javax.persistence.EntityNotFoundException;
 
 @RestController
 @RequestMapping("/transactions")
@@ -27,11 +33,15 @@ public class TransactionsController {
 	
 	@GetMapping
 	public List<TransactionsEntity> getAllTransactions() {
+		if (tr.findAll().size() <= 0) {
+			throw new SpringBankingServerException("Could not retrieve transactions from server");
+		}
 		return tr.findAll();
 	}
 	
 	@GetMapping("/id/{id}")
 	public TransactionsEntity getOneTransaction(@PathVariable Integer id) {
+		System.out.println(tr.getById(id));
 		return tr.getById(id);
 	}
 	
@@ -61,18 +71,21 @@ public class TransactionsController {
 		return ts.getAllTransactionsByAccountID(a.getAccountID());
 	}
 	
-//	@PostMapping("transfer")
-//	public List<TransactionsEntity> getAllDepositTransactions() {
+//	@PostMapping("/transfer")
+//	public List<TransactionsEntity> getAllDepositTransactions(@RequestBody List<TransactionsEntity> transactionsEntity) {
+//		ts.transferAccounts(transactionsEntity);
 //		return ts.getAllTransactions();
 //	}
 	
 	@DeleteMapping("/id/{id}")
 	public String deleteOneTransaction(@PathVariable Integer id) {
-		tr.deleteById(id);
-		return "done";
+		if (id == 0) {
+			throw new SpringBankingAPIException("Please enter an transaction id greater than 0");
+		}
+		if (tr.findAllByAccountID(id).isEmpty()) {
+			throw new NoSuchElementException("Could not find transaction with id:" + " " + id);
+		}
+		 tr.deleteById(id);
+		 return "done";
 	}
-	
-	
-	
-
 }
